pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "owly-ai"
version = "0.1.1"
description = "Streaming-first LLM runtime for real-time systems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Owly Contributors" }
]
keywords = ["llm", "streaming", "async", "realtime", "voice", "chat", "agents"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
]
dependencies = ["openai", "google-genai>=1.0", "anthropic"]

[project.urls]
Homepage = "https://github.com/amrawlabs/owly"
Repository = "https://github.com/amrawlabs/owly"
Issues = "https://github.com/amrawlabs/owly/issues"

[tool.hatch.build.targets.wheel]
packages = ["owly_ai"]
